Passing 3 tests. fixed gitignore
I am currently passing the first three tests. Considering the fact that i
do not think the first two tests are actually testing anything it is not
much or an accomplishment. Both delete matches and delete players need to
be fixed so they add players and matches, check the count and then delete
all of them again.